move storage trait into a separate file
