+++ src/building_blocks/memtable/mod.rs
@@ -1,6 +1,7 @@
 mod memtable;
 mod memtable_entry;
+mod storage_trait;
 
 pub use memtable::Memtable;
-pub use memtable::StorageCRUD;
 pub use memtable_entry::MemtableEntry;
+pub use storage_trait::StorageCRUD;
